Ajout d'un manager pour la generation de clusters et des detecteurs
