--- src/fr/jgetmove/jgetmove/solver/ISolver.java
@@ -1,5 +1,0 @@
-package fr.jgetmove.jgetmove.solver;
-
-public interface ISolver {
-
-}
